R6 - Cantidades por Kilo/Metro con decimales:
- INTEGER -> NUMERIC(12,3) en guia_items, guia_item_partidas, etiquetas,
  notas_salida_detalle (cantidad + devuelto_cantidad), inventario,
  movimientos. unidades_medida.permite_decimal (KG/M/M2/L).
- Validadores decimales (esCantidadPositiva, hasta 3 decimales); casts
  ::int -> ::float8 en las agregaciones que consume el front.
- Devolucion parcial: cobra solo lo consumido -> notas_salida_detalle
  .cantidad_consumida / .total_consumido; el codigo reingresado marca
  el sobrante.

R4 - ID de agrupacion, observaciones y codigo de impresion:
- guias.observaciones (antes se imprimia sin que la columna existiera);
  guia_items.id_agrupador (repetible), observaciones, codigo_impresion.
- codigo_impresion es solo la etiqueta a imprimir: NO toca etiquetas.codigo
  ni el codigo de barras del sticker.
- Alta y edicion de guia con esos campos; modal "Revisar antes de imprimir".
- Nota de Ingreso impresa: 4 columnas del talonario fisico + celda DETALLE
  enriquecida (ITEM n+ID por grupo, ubicacion, motivo/maquina, codigo a
  imprimir, observacion de linea); lineas ordenadas por id_agrupador.

Instalador 1.5.0 -> 1.6.0. Detalle por paso en NOTAS_PROYECTO_SIBARITA.txt
(seccion 22). Pendiente aparte: formato "NOTA DE INGRESO - DESUSO".

+++ backend/src/routes/dashboard.js
@@ -11,7 +11,7 @@ router.get('/fase2', verificarToken, async (req, res) => {
       pool.query('SELECT COUNT(*)::int as total FROM etiquetas'),
       pool.query(`SELECT COUNT(*)::int as total FROM notas_salida WHERE estado IN ('PENDIENTE', 'EN_APROBACION')`),
       pool.query(`
-        SELECT p.nombre as producto_nombre, SUM(gi.cantidad)::int as total_movido
+        SELECT p.nombre as producto_nombre, SUM(gi.cantidad)::float8 as total_movido
         FROM guia_items gi
         JOIN productos p ON gi.producto_id = p.id
         WHERE gi.tipo = 'PRODUCTO'
